[fix] 모임 작성자 자동참여,클릭불가

+++ src/main/java/sparta/seed/service/CommunityService.java
@@ -71,8 +71,8 @@ public class CommunityService {
               .title(community.getTitle())
               .isRecruitment(getDateStatus(community).equals("before"))
               .participantsCnt(community.getParticipantsList().size())
-              .currentPercent(((double) community.getParticipantsList().size() / (double) community.getLimitParticipants()) * 100) // 참여인원 퍼센트
-              .successPercent((Double.valueOf(certifiedProof) / (double) community.getLimitScore()) * 100) // 아직 미적용
+              .currentPercent(((double) community.getParticipantsList().size() / (double) community.getLimitParticipants()) * 100)
+              .successPercent((Double.valueOf(certifiedProof) / (double) community.getLimitScore()) * 100)
               .isWriter(userDetails != null && community.getMemberId().equals(userDetails.getId()))
               .dateStatus(getDateStatus(community))
               .build());
@@ -86,21 +86,10 @@ public class CommunityService {
   public ResponseEntity<Community> createCommunity(CommunityRequestDto requestDto, List<MultipartFile> multipartFile, UserDetailsImpl userDetails) throws IOException {
     Long loginUserId = userDetails.getId();
     String nickname = userDetails.getNickname();
-    List<Img> imgList = new ArrayList<>();
     if (multipartFile != null) {
-      Community community = Community.builder()
-              .title(requestDto.getTitle())
-              .content(requestDto.getContent())
-              .isSecret(requestDto.isSecret())
-              .password(requestDto.getPassword())
-              .memberId(loginUserId)
-              .nickname(nickname)
-              .startDate(requestDto.getStartDate())
-              .endDate(requestDto.getEndDate())
-              .limitParticipants(requestDto.getLimitParticipants())
-              .limitScore(requestDto.getLimitScore())
-              .build();
-
+      Community community = getCommunity(requestDto, loginUserId, nickname);
+      Participants groupLeader = getGroupLeader(loginUserId, nickname, community);
+      List<Img> imgList = new ArrayList<>();
       for (MultipartFile file : multipartFile) {
         S3Dto upload = s3Uploader.upload(file);
         Img findImage = Img.builder()
@@ -110,11 +99,22 @@ public class CommunityService {
                 .build();
         imgList.add(findImage);
         imgRepository.save(findImage);
+
       }
       communityRepository.save(community);
+      participantsRepository.save(groupLeader);
       return ResponseEntity.ok().body(community);
     }
-    Community community = Community.builder()
+    Community community = getCommunity(requestDto, loginUserId, nickname);
+    Participants groupLeader = getGroupLeader(loginUserId, nickname, community);
+    communityRepository.save(community);
+    participantsRepository.save(groupLeader);
+    return ResponseEntity.ok().body(community);
+
+  }
+
+  private Community getCommunity(CommunityRequestDto requestDto, Long loginUserId, String nickname) {
+    return Community.builder()
             .title(requestDto.getTitle())
             .content(requestDto.getContent())
             .isSecret(requestDto.isSecret())
@@ -126,9 +126,14 @@ public class CommunityService {
             .limitParticipants(requestDto.getLimitParticipants())
             .limitScore(requestDto.getLimitScore())
             .build();
-    communityRepository.save(community);
-    return ResponseEntity.ok().body(community);
+  }
 
+  private Participants getGroupLeader(Long loginUserId, String nickname, Community community) {
+    return Participants.builder()
+            .nickname(nickname)
+            .memberId(loginUserId)
+            .community(community)
+            .build();
   }
 
 
@@ -191,12 +196,16 @@ public class CommunityService {
    * 그룹미션 참여 , 취소 하기
    */
   @Transactional
-  public ResponseEntity<Boolean> joinMission(Long id, UserDetailsImpl userDetails) {
+  public ResponseEntity<Boolean> joinMission(Long id, UserDetailsImpl userDetails) throws Exception {
     Optional<Community> community = communityRepository.findById(id);
     Long loginUserId = userDetails.getId();
     String nickname = userDetails.getNickname();
     long limitParticipantCount = community.get().getLimitParticipants();
     int participantSize = community.get().getParticipantsList().size();
+    if (community.get().getMemberId().equals(userDetails.getId())) {
+      throw new Exception("작성자는 누를수 없습니다");
+    }
+
     if (participantsRepository.existsByCommunityAndMemberId(community.get(), loginUserId) || participantSize >= limitParticipantCount) {
       participantsRepository.deleteByMemberId(loginUserId);
       return ResponseEntity.ok().body(false);
@@ -221,7 +230,7 @@ public class CommunityService {
   }
 
 
-  private String  getDateStatus(Community community) throws ParseException {
+  private String getDateStatus(Community community) throws ParseException {
     return dateUtil.dateStatus(community.getStartDate(), community.getEndDate());
   }
 }
